#130 renamed region_polygon_file.py to polygon_file.py
